Implemented review comments from review 2267 Related to Issue [WTI-1298]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@12720 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: e35a56a0daed4a728dc120c2054f1318de189d08
Former-commit-id: 35409add7a8a1a7fdbf85242db5c5f3593812733

+++ Application/Ringtoets/src/Application.Ringtoets.Storage/ColorConversionExtensions.cs
@@ -31,15 +31,15 @@ namespace Application.Ringtoets.Storage
     internal static class ColorConversionExtensions
     {
         /// <summary>
-        /// Convert <paramref name="value"/> to its 32-bit ARGB value.
+        /// Convert <paramref name="value"/> to its 64-bit ARGB value.
         /// </summary>
         /// <param name="value">The value to convert.</param>
-        /// <returns>The 32-bit ARGB value, or null when the
-		/// color is empty.</returns>
-        public static int? ToInt32(this Color value)
+        /// <returns>The 64-bit ARGB value, or null when the
+        /// color is empty.</returns>
+        public static long? ToInt64(this Color value)
         {
             return value.IsEmpty
-                       ? (int?) null
+                       ? (long?) null
                        : value.ToArgb();
         }
